Minor updates to address comments
